Added Time
Time now displays in two different selects, one for the hour and one
for the minutes. The user will have to pick the hour for the
notifications to appear.
Functionality is stopped for now, it only displays the time, but
threads won't start yet.

*All notifications will appear at the same time now. Later
implementation will give each notification a different time*